BUG: Fix searchsorted and CheckFromAny byte-swapping logic
* BUG: Fix searchsorted and CheckFromAny byte-swapping logic

This closes gh-28190 and fixes another issue in the initial code
that triggered the regression.

Note that we may still want to avoid this, since this does lead to
constructing (view compatible) structured dtypes unnecessarily here.

It would also compactify the dtype.  For building unnecessary dtypes,
the better solution may be to just introduce a "canonical" flag to
the dtypes (now that we have the space).

* STY: Adopt code comment suggestions