MAINT: fix typo in normal distribution functions docstrings
This commit replaces an incorrect expected result in the docstring of
the numpy.random.normal() and numpy.random.RandomState.normal() versions
of the normal distribution function by the correct value; namely, the
standard deviation of the normal distribution should be (closely equal
to) the sigma input parameter.

Note that this typo was partially fixed in commit f2faf84, but only for
the random.Generator.normal() version of the function.