API: Always allow ``sorted=False`` and make a note about it
User intent for `sorted=False` is very clear, so I think we can
really just allow it.
However, note in the docs that it could change.  Also add a bigger
note in the `unique_*` functions that still always return a sorted
result but would change in the future (whenever `sorted=False` does
work).