API,ENH: Allow forcing an array result in ufuncs
As of now a WIP, because it would be nice to see what `_methods.py`.

The API change is to allow `out=...`, for multi-argument outputs it
is currently `out=(..., ...)`, although I guess we should allow
short-handing that (nobody needs to pick it per argument).

There is some discussion in gh-14489 about this.  I still think `...`
is the best, because `out=np.ndarray` doesn't generalize for
`__array_ufunc__`, and yes, I like `...` because it also indicates an
array return in indexing.  (`arr[i, ...]` always returns an array.)

---

The ufunc part works (not sure if 100% complete), but the `_methods.py`
changes do not (and change behavior subtlely, if possibly correctly).
The actual change is trivial (after I had refactored this a while ago).

I think we should do this, but the fact that the `_methods.py` changes
are not trivial makes me think there may be some idea missing.

It may be that the solution is for these functions to use conversion
helper once at the start and then convert to scalars based on that at the
end though (rather than dragging around subclasses).