API: Allow forcing an array result in ufuncs (#28576)
* API,ENH: Allow forcing an array result in ufuncs

This allows using `out=...` for ufuncs (including reductions) to enforce an array result.
With it, the result will never be converted to a scalar, but subclasses are preserved
normally.
(A subclass might still misbehave and convert to a scalar anyway, but it would be a bug there.)

There is some discussion in gh-14489 about this.  I still think `...`
is the best, because `out=np.ndarray` doesn't generalize for
`__array_ufunc__`, and yes, I like `...` because it also indicates an
array return in indexing.  (`arr[i, ...]` always returns an array.)

It uses this in a few places, although there are many more, e.g. `_methods.py`
should probably use it, but it is very subtle there.
(Also because some paths are probably subtly wrong for object dtype.)

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>
Co-authored-by: Marten H. van Kerkwijk <mhvk@astro.utoronto.ca>
Co-authored-by: Nathan Goldbaum <nathan.goldbaum@gmail.com>