WIP, MAINT: scalar float str match array
* Fixes gh-28679, but mostly an initial demo for now...

* This patch sends scalar floating point string representations
through the exact dragon representation path rather than the unique
path for the reasons outlined in the matching ticket. It does not
concern itself with whether the final `integer.0` is printed or excluded
via `integer.`. The main argument in favor is that it is pretty weird
to follow a different path for 0-D arrays vs. multi-element arrays.
The main argument against is likely the potential shenanigans this
kind of large-scale reprsentation change could entail downstream
if folks already have shims around it and if the rounded representations
are not truly bugs but just different design decisions.

* A `hypothesis` test was added that fails before and passes after
the switch to the exact scalar flot representations. `hypothesis`
was chosen because the idiosyncrasies of asymmetric IEEE floating
point number lines are well known and trying to test them with
a few hard-coded manual cases seems like a recipe for missing something.

* I think the NumPy team tends not to like partial fixes, so if we do
this we may also want to apply the same decision to scientific
notation representations, which I have not touched here yet, but the
changes here should suffice to provide enough of a preview for decision
making on the matter.

* The `stringdtype` testing changes are rather egregious, so that's
probably enough of a sample of potential drawbacks to weigh the
design considerations for feedback. More gruesome changes would
likely be needed for other tests based on local experiments--I see
some other failures related to array/scalar/string round-tripping
guarantees that would need careful thought and that are broken
by this set of adjustments.

[ci skip] [skip ci]