BUG: fix ``np.vectorize`` for object dtype (#28624)
With `dtype=object` results, the `asarray()` call previously lead to incorrect results when the inputs were scalars.  The new `out=...` provides a way to ensure the result is always an array.
(Unlike ufuncs `vectorize` always returns arrays also for 0-D results, which is maybe weird but not a change here.)

Closes #21274