BUG: Fix #27881 Adding tests for numpy.roots and numpy.polynomial.polynomial.polyroots
The following tests provide a minimum precision expected by the functions. In the case of my change, they allow my change but not the previous version. You'll find that the difference of precision between the two versions vary according to the root values but in all cases, this change either increases precision or maintains the previous result.