BUG: Use unrotated companion matrix in polynomial.polyroots. (#28821)
* BUG: Fix #27881 inconsistent behavior between numpy.roots and numpy.polynomial.polynomial.polyroots

Both functions use numpy.linalg.eigvals, but while roots gives in argument the polynomial's companion matrix unchanged, polyroots rotates it. Though in theory this rotation shouldn't change anything, in terms of numerical calcuations, eigvals gives significantly different results. This commit removes the rotation as an easy fix to the inconsistency.

This strange behavior by eigvals is however a bug. I did some research on it, which you can find on the issue.

* BUG: Fix #27881 Adding tests for numpy.roots and numpy.polynomial.polynomial.polyroots

The following tests provide a minimum precision expected by the functions. In the case of my change, they allow my change but not the previous version. You'll find that the difference of precision between the two versions vary according to the root values but in all cases, this change either increases precision or maintains the previous result.

* BUG: Fix #27881 Adding tests for numpy.roots and numpy.polynomial.polynomial.polyroots

Fixing tests

* BUG: Fix #27881 Adding tests for numpy.roots and numpy.polynomial.polynomial.polyroots

Fixing tests. It seems that the way numpy.roots creates it's matrice companion augments the numerical error compared to the polycompanion function, which is why I had to reduce the expected precision.

* Reverting change from commit 6703b91a22f7f0d1e98f565efcb659d8369b4952 in matlib.pyi