TST: Prevent import error when tests are not included in the wheel
It is possible to build numpy without test directories using install
tags. For example, passing the tags such as

install-args=--tags=runtime,python-runtime,devel

would exclude the test directories. However, even when the test
directories are excluded, numpy.testing is included in the wheel and one
of the file in numpy.testing tries to import the file from the test
directory, which causes an error.