MAINT: Fix for segfaults with GCC 15
GCC 15 generates code that segfaults on newer hardware when running
tests. See gh-28991 for details. A fix is to always require alignment,
which can be done by setting `NPY_ALIGNMENT_REQUIRED = 1`.

We retain `NPY_ALIGNMENT_REQUIRED` here for downstream compatibility in
the 2.3.x release series, but do remove its use in our own code. The
associated macro `NPY_USE_UNALIGNED_ACCESS` could also be removed as it
is always 0, but that is left for another PR.