Fix NEON_FP16 check for MSVC
The cast in `numpy/distutils/checks/cpu_neon_fp16.c` didn't compile correctly in MSVC, which lead to `NEON` reporting as unsupported due to it encompassing `NEON_FP16` in `meson_cpu/arm/meson.build`.

I'm not an expert, but it appears `fp16`/`fp16fml`/`dotprod` are not supported in VS2022:
https://learn.microsoft.com/en-us/cpp/build/reference/feature-arm64?view=msvc-170

Similarly, I don't think MSVC supports `float16` as a type yet, so I've enabled emulation in the strided loop code.