BUG: linspace: Round towards `start` with integer dtype.
The PR #16841 implemented rounding towards `-inf` for integer dtypes.
While this is for sure better than the old towards `0` approach, it does
not produce good results for descending intervals with `start` > `stop`.

It violates (at least my) expected behavior of `-linspace(A, B) == linspace(-A, -B)`.
And it violates the documentation of `endpoint` as `stop` can be included.

However I'm not sure if these conflicts are worth to break the behavior of `linspace`,
even for the edge cases of `dtype=integers` and `start > stop`.

Related to #16813