BUG: Fix np.testing utils failing for masked scalar vs. scalar (#29317)
TestArrayEqual.test_masked_scalar now passes.

This case regressed since 73151451437 (merged in #12119) due to:
- `<masked scalar> == <scalar>` returning np.ma.masked (not a 0-dim
   masked bool array), followed by
- `np.bool(np.ma.masked)` unintentionally converting it to np._False

There are a few ways to resolve this; I went with testing the
comparison result with `isinstance(bool)` to check if a conversion to
array is necessary, which is the same approach already taken in
assert_array_compare after evaluating `comparison(x, y)`.