BUG: Fix np.testing utils failing for masked scalar vs. scalar (#29317)
TestArrayEqual.test_masked_scalar now passes.

This case regressed since 73151451437 (merged in #12119) due to:
- `<masked scalar> == <scalar>` returning np.ma.masked (not a 0-dim
   masked bool array), followed by
- `np.bool(np.ma.masked)` unintentionally converting it to np._False

Note on the modified comment: Confusingly, "isinstance(..., bool)
checks" in the previous wording actually incorrectly referred to the
ones towards the end of the function, which are not actually related
to __eq__'s behavior but to the possibility of `func` returning a bool.