ENH: Improve error message in numpy.testing.assert_array_compare (#29112)
Improve the error message in numpy.testing.assert_array_compare to show the position of differing elements.

Currently, if there is an assertion error in functions like numpy.testing.assert_allclose, both arrays are printed. Typically the arrays are so large that they are truncated before being printed, making it hard to spot the differences. I suggest to print the indices of the differing elements as well as the corresponding values. To avoid excessive output, the number of printed out differences is limited to five.

* MAINT: Different error messages for different numbers of mismatches