API,BUG: Fix scalar handling in array-interface allowing NULL pointers (#29338)
This fixes the scalar handling in the array-interface and also a bug
in the shape handling error path.

In theory it breaks code that used NULL to trigger the scalar path,
but this path was:
1. Completely undocumented, and
2. correctly triggered by *ommitting* the data field instead.

I didn't remove/deprecate the scalar path in this PR, maybe we should.
But, I do think we can ignore that theoretical use-case since it is
nonsensical.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>