BLD: provide explicit control over cpu-baseline detection
Add a new `cpu-baseline-detect` feature flag that can be used to more
precisely control the use of CPU baseline detection.  This can be used
by packages to more precisely control used SIMD code independently
of compiler flags specified.  The option follows typical feature
semantics -- with `auto` preserving the current behavior of enabling
when relevant compiler flags are found, `enabled` forcing it on
based on the implicit compiler defaults, and `disabled` forcing it off.

Signed-off-by: Michał Górny <mgorny@quansight.com>