MAINT: Replace setting of array shape by reshape operation (#29314)
This PR we replace some explicit setting of the shape of an array in favor of using reshape. Some cases are left out (e.g. masked arrays, tests to check the .shape setting works).

This is because explicit shape setting is generally unsafe mutation that we would like to avoid/get rid of.