MAINT: Eliminate ambiguous order of evaluation in the ratios of random variates. (#29598)
In C, in an expression such as f1()/f2(), the order of evaluation of f1()
and f2() is unspecified.  When such a ratio was used in the C functions
random_f and random_standard_cauchy, it meant the streams of random
variates for the f and standard_cauchy distributions depended on how the
compiler chose to order the evaluation.  This could result in different
streams of variates when numpy is compiled with different compilers.

By evaluating the numerator and denominator in separate statements, the
ambiguity is eliminated.


Co-authored-by: Joren Hammudoglu <jhammudoglu@gmail.com>