ENH: A extended sorting APIs
This PR adds new sorting APIs that can be used for extending our current
sort types. The new APIs dispense with the *kind* argument in the legacy
interface with three new boolean arguments:

- stable -- whether the sort should stable, default is False,
- descending -- whether to sort in descending order, default is False,
- nanfirst -- whether NaNs sort to the last or first, default is False.

The main difference is that sorts are now chosen by functionality
instead of by algorithm.

The functions added are:

- (API)PyArray_SortExt,
- (ndarray method)sortext,
- (function)sortext,
- (API)PyArray_ArgSortExt,
- (ndarray method)argsortext,
- (function)argsortext.

Both the legacy and new methods and functions call through the
API functions, so those are the two functions that need changes
depending on how we want to implement adding new sort algorithms.
The legacy APIs were left for downstream compatibility.

Thought was given to just modifying the legacy methods and functions to
optionally use the new arguments, but documenting that made such
a complicated mess that I decided it was unusable in practice. It would
confuse the heck out of users.

One change that has happened under the covers is that heapsort is no
longer publically accessible. It is used by introsort, so the code
remains, but in practice its only virtue was that it was both inplace
and always ``nlog(n)``. Introsort has the same virtue and is usually
faster, so it is used in its place. I doubt anyone will notice.

I note that the partioning/selection functions have not been modified,
we may want to do that a some point.