BUG: universal infinity min/max filler for objects
Adds _UniversalInfinity object which (preferentially) compares greater
than all other objects other than itself.  This is used as the default
min_filler.  Previously, min filled masked values with the string '?',
which caused a TypeError when the array contained objects that could not
be compared with str.
Likewise, negative universal infinity is used as the max_filler.
See #9103.