TST: test #9103 reproducer specifically
NOTE: The reproducer given in #9103, although reported as producing an
AttributeError, actually produced a TypeError.  The AttributeError was a
secondary issue which occurs when there is no mask.  Because there were
2 bugs associated with the issue (one of was unnoticed in my previous
attempt to fix the issue), we test the reproducer specifically just as
confirmation that the issue is fully resolved and I'm not missing
anything else.
This test *should* be redundant with the added test_minmax_object_dtype
tests, but it's not exactly the same.