BUG: Fix assert in nditer buffer setup
When using a buffered iteration, there was a new assert to check
that strides were set up nicely to ensure that we are not missing
dimension coalescing.

However, when the strides are set up to track an index, then they
were still set up with a 0 for length 1.
I _think_ that is just unnecessary (i.e. the assert is correct to
point it out).  But, let's do this first for backporting at least.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>