DOC: clarify numpy.asarray, numpy.asanyarray, numpy.asarray_chkfinite parameter 'order'  (#29655)
* DOC: clarified numpy.asarray parameter 'order'

Changed description to make explicit that setting 'A' requires
user to set 'copy' to 'True' to get the described behavior.
Made explicit that 'A' has two cases where it acts as
equivalent to setting 'F'. See 28247.

* DOC: clarified numpy.asarray parameter 'order'

See 28247. Changed description of 'order' to make
the behavior of 'A' more explicit and how to cause it to
preserve non-contiguous Fortran-order. Changed description
of 'C' and 'F' to clarify that output will be contiguous.

* DOC: clarified numpy.asarray() parameter 'order'

See 28247. To differentiate between 'A' and 'K', made explicit
that 'A' will not ensure the output is contiguous. 'K' has more
complex behavior related to flags and stride detection, but will
try to keep the output as close as possible to the input.

* DOC: clarify numpy.asarray parameter 'order'

See 28247. Made explicit that 'C' and 'F' can copy input to output,
'A' does not ensure contiguous output.

* DOC numpy.asarray parameter 'order' description added line break

* DOC: Additional request to change identical text blocks.

* DOC: fixed linebreak position