ENH: Add fast path in ufuncs for numerical scalars.
This adds a fast check for numerical scalars, improving the speed by
about a factor of 6, with essentially no cost for arrays.  At the
moment, just for single-input, single-output ufuncs, with the
assumption that the output has the same type as the input (if that
fails, the regular path is taken).