ENH, FEAT: Reorganize finfo and add new constant slot (#29836)
Introduce `NPY_DT_get_constant` as a slot and a corresponding function that fills in a single element pointer for an arbitrary (corresponding) dtype/descriptor.
Since we do want some integer values for `finfo`, some values are set to always fill in an `npy_intp` value instead.

The slot assumes that it will be used with the GIL held and on uninitialized data, since I suspect that this is the typical use-case (i.e. it is unlikely that we need to fetch a constant deep during a calculation, especially in a context where the correct type/value isn't known anyway).

This completely re-organizes and simplifies `finfo` by moving all definitions to C.

Co-authored-by: Sebastian Berg <sebastianb@nvidia.com>