ENH, PERF: allocate dims and strides on array object by default.
Allocate memory for the dimensions and strides as part of the object,
to avoid the overhead of an extra allocation.  This will only become
useless memory if the array shape is set (now deprecated) or the array
is resized in-place (frowned upon).

Note that the above is done only for standard arrays, though I think
in principle it could be extended to subttypes with a suitable check.

Also, allocating space for data is done only for the standard
allocator.

Timings
```
%timeit np.empty(())
133->127 ns

a = 1.6
%timeit np.array(a)
151->137 ns

%timeit np.add(a, a)
581->557 ns

b = np.array(a)
%timeit np.add(b, b)
293->279 ns
```