ENH, PERF: allocate memory as part of the array object for scalars
Allocate memory for small amounts of data as part of the object, to
avoid the overhead of the allocation of tracked memory.

Like for the dimensions and strides, this is done only for standard
arrays.  Furthermore, allocating space for data is done only for the
standard allocator.

Timings relative to having dimensions, strides and data not on the
object:
```
%timeit np.empty(())
133->73 ns

a = 1.6
%timeit np.array(a)
151->92 ns

%timeit np.add(a, a)
581->431 ns

b = np.array(a)
%timeit np.add(b, b)
293->248 ns
```