MAINT: Rewrite setitem to use the new API (mostly)
This has a mild speed advantage since `PyArray_Pack` had this hack
in place to ensure that we have an "array" to pass on.  This hack
is now only used in extreme cases (i.e. user dtypes).

This gives a very slight speed advantage, but mostly, I like to
change this.

It would be nice to change `getitem` as well, but `getitem` is
trickier due to structured dtypes.

I wanted to first just remove `->setitem` but one (brutal!) test
failed and I reconsidered things thinking that someone might be
using `descr->setitem` without an array which would be broken
without keeping an explicit definition around.

The 0-D unpacking path should be unimportant these days (it only
applies to _nested_ arrays so the tests don't even notice it)
but I kept it around for now anyway.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>