ENH: Set DLPack tensor `shape` and `strides` to NULL iff `ndim == 0`. (#29872)
This assumes a slight shift in the dlpack interpretation:

https://github.com/dmlc/dlpack/issues/177

---

DLPack allows, but does not require, DLTensor.strides to be NULL if PyArray_IS_C_CONTIGUOUS.
Previously, in this case, NumPy set the strides pointer to NULL.
Now, NumPy will supply a valid pointer to correctly initialized strides when ndim > 0.

This change can improve performance for downstream users. For example, nanobind can be used to import a NumPy array using the DLPack protocol to make the array available to C++ code. If DLTensor.strides is NULL, nanobind allocates and initializes a strides array. This redundant work is now avoided since NumPy provides a pointer to the strides, which it allocated and initialized anyway.

If ndim == 0, this PR sets both shape and strides to NULL.
Previously, strides was was set to NULL (since PyArray_IS_C_CONTIGUOUS), but shape pointed past the end of allocated memory. For security and debugability, NULL is preferred.