MAINT: replace use of asanyarray with ... to keep arrays (#29951)
With gh-28576, it has become possible to ensure that when calling a ufunc, the output is guaranteed to be an array.

This PR uses that to replace np.asanyarray calls in some functions. I'm not sure this is complete -- these were just parts of code for which test failed if np.anyarray(scalar) is made to return a read-only array (see gh-29876).

I do think there are all small improvements, though.