Do not use negative values in argument of Tensor.<type>Max()
The unit tests involving `<type>TestCase.testMaxWrongDim` do not work for
`<type>` being one of `uchar`, `ushort`, `uint`, `ulong`, and
`ulongLong` (which translate to `uint8`, `uint16`, `uint32`, `uint64`, and
`uint64`, respectively). Instead of raising the expected `TypeError`, the
unit tests raise `OverflowError`, with the error message “Python integer -1
out of bounds for <type>”. This is fixed by replacing the negative values
in the argument of the `max` function by positive ones.