MAINT: undo change to `fromstring` text signature for 2.4.0
This reverts a very small part of gh-30147, because it broken Pythran
(see pythran#2368). While a new Pythran 0.18.1 release is available,
that doesn't help for the SciPy 1.15.x releases which allow
numpy<2.5 and pythran<0.18, see
https://github.com/scipy/scipy/blob/maintenance/1.15.x/pyproject.toml#L30-L41
So introducing this signature change only for 2.5.0 is much better.

In addition, doing so doesn't leave us only a very narrow window with
compatible NumPy and Pythran releases which would otherwise bite distro
packagers as well.