MAINT,API: Introduce __numpy_dtype__ and fix dtype attribute recursion (#30179)
* MAINT,API: Introduce __numpy_dtype__ and fix dtype attribute recursion

This is a lot more complex then I expected. We, correctly, deprecated
`.dtype` attribute lookup recursion, however...
The code still had a `try/except` that still recursed and that try
except carried meaning at least in a weird path of:
```
class mydtype(np.void):
    dtype = ...
```

Now does that make sense? Maybe not... we also fall back to object
in some paths which should have been broken when a dtype attribute
existed on a type but it is a property/descriptor.

So, this removes the recursion, but adds a check for `__get__` to
filter out those cases, this is something we successfully did for
other protocols `__array__`, `__array_interface__`, etc.

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>

* DOC: Add release note snippet

* DOC: Make release note more precise, array-likes don't need the new attr

---------

Signed-off-by: Sebastian Berg <sebastianb@nvidia.com>