DOC: Improve cross-links in thread safety documentation (#30373)
* DOC: Improve cross-links in thread safety documentation

Those are mostly cosmetic changes and extra cross-links to improve the
documentation. In particular I'd like to use this page as an example of
good top-level documentation for threading/freethreading in the
py-threading-guide https://py-free-threading.github.io/

This also does a number of updates to the misc.rst file which was
basically too indented, and missing a bunch of links. I think many of
the info misc.rst on how to interop with C/C++ could be removed, and likely
all the information on seterr/geterr moved to routines.err.rst but I don't
want to include this moves as part of this PR as it's likely more
subject to discussion.

There are minor updates in conf.py to avoid 1 sphinx warnings, and I
know the thread safety backlink in geterr/seterr/...etc, are a bit overkill,
but as numpy will be used as an example and in general is referred to as
one of the core scientific Python project I thinks it's OK go go a bit
further than usual.

[skip actions]
[skip cirrus]

* typo