DOC: Make the issue template clearer about what "context" means
I like the "context", but more often than not, the context is not
provided or filled with rather meaningless repititions of how it
*might* affect someone.

So I think it may be good to be more blunt about what we at least
I would like to see here (i.e. concrete impact to the reporter).