BUG: Fix leak in flat assignment iterator
Both the value array and iterator were missing the DECREF.
Since the iterator is a temporary object, I don't really think it is
particularly helpful to try to test.
Rather, it would be good to try to test with ASAN semi-regularly
(or I use `pytest-valgrind` again, which works great but is just very
very slow).

Closes gh-30508